refactor(accent): split into api/accent/ package (no behavior change)
Pure structural split of the monolithic api/accent_marker.py (493L) +
api/furigana_marker.py (152L) into a focused package layout. Zero
behavior change — verified byte-identical /MarkAccent/ and
/MarkFurigana/ responses before and after.

Layout:
  api/accent/
    __init__.py     re-exports accent_router, furigana_router
    README.md       pipeline + accent_marking_type semantics
    models.py       Request, ErrorInfo, WordResult, WordAccentResult,
                    AccentInfo, FuriganaResponse, AccentResponse
    furigana.py     Yahoo Furigana HTTP client (data layer)
    ojad.py         OJAD scrape (HTTP + BS4 parse)
    align.py        align_accent + numeric_pattern + punctuation_marks
                    + skip_marks + clean_query + is_kana_or_kanji
    pipeline.py     MarkAccent orchestrator (extracted from inline
                    mark_accent handler)
    routes.py       FastAPI routers + thin endpoint handlers

main.py: drop `accent_marker` and `furigana_marker` from the
`from api import ...` line; add `from api.accent import accent_router,
furigana_router`; rename the two `include_router` calls accordingly.

README.md documents the data flow (mermaid), AccentInfo
accent_marking_type semantics (0=LOW, 1=HIGH, 2=FALL), heiban /
atamadaka / nakadaka / odaka detection rules, file responsibilities,
and the alignment algorithm. Forms the foundation for #47/#51 to
extend with override / patch / postprocess layers.

Verification:
  uv run ruff check api/accent/ main.py        # all passed
  uv run mypy api/accent/ main.py              # 8 files, no issues
  POST /api/MarkAccent/  → byte-identical to pre-refactor
  POST /api/MarkFurigana/ → byte-identical to pre-refactor
