--- api/furigana_marker.py
@@ -1,152 +1,0 @@
-"""
-An API that mark furigana of given query text
-"""
-
-from __future__ import annotations
-
-import httpx
-from fastapi import APIRouter, Depends
-from pydantic import BaseModel, Field
-
-from api.dependencies import get_http_client
-from config.settings import YAHOO_API_KEY
-
-tags_metadata = [
-    {
-        "name": "MarkFurigana",
-        "description": "Mark furigana of given text",
-    },
-]
-
-
-class Request(BaseModel):
-    """Class representing a request object"""
-
-    text: str = Field(description="The text to query")
-
-
-class ErrorInfo(BaseModel):
-    """Class representing an error information"""
-
-    code: int = Field(description="The error code that follows JSON-RPC 2.0")
-    message: str = Field(
-        description="The error message that describe the details of an error"
-    )
-
-
-class WordResult(BaseModel):
-    """Class representing a word result object"""
-
-    furigana: str = Field(description="Furigana of given kana and kanji")
-    surface: str = Field(description="The (partial of) original query text")
-    subword: list[WordResult] = Field(
-        default_factory=list,
-        description="""A list contains more details when a \
-        word contains both kanji and kana. Each elements in \
-        subword follow the same schema as this parent object \
-        (containing `furigana`, `surface`, and `subword`).""",
-    )
-
-
-class Response(BaseModel):
-    """Class representing a response object"""
-
-    status: int = Field(
-        default=200, description="Status code of response align with RFC 9110"
-    )
-    result: list[WordResult] | None = Field(
-        description="A list contains marked results"
-    )
-    error: ErrorInfo | None = Field(
-        default=None,
-        description="An object that describe the details of an error when occur",
-    )
-
-
-router = APIRouter()
-
-url = "https://jlp.yahooapis.jp/FuriganaService/V2/furigana"
-
-
-@router.post("/MarkFurigana/", tags=["MarkFurigana"], response_model=Response)
-async def mark_furigana(
-    request: Request,
-    client: httpx.AsyncClient = Depends(get_http_client),
-) -> Response:
-    """Receive POST request, return a Response object"""
-
-    # 輸入
-    headers = {
-        "Content-Type": "application/json",
-        "User-Agent": f"Yahoo AppID: {YAHOO_API_KEY}",
-    }
-
-    data = {
-        "id": "1234-1",
-        "jsonrpc": "2.0",
-        "method": "jlp.furiganaservice.furigana",
-        "params": {"q": request.text, "grade": 1},
-    }
-
-    try:
-        response = await client.post(url, headers=headers, json=data)
-
-    except httpx.TimeoutException:
-        return Response(
-            status=408,
-            result=None,
-            error=ErrorInfo(code=408, message="Yahoo API request timed out"),
-        )
-
-    except httpx.HTTPError as e:
-        return Response(
-            status=500,
-            result=None,
-            error=ErrorInfo(code=500, message=f"HTTP error: {str(e)}"),
-        )
-
-    if response.status_code != 200:
-        return Response(
-            status=response.status_code,
-            result=None,
-            error=ErrorInfo(
-                code=response.status_code,
-                message=f"Yahoo API request failed with status {response.status_code}",
-            ),
-        )
-
-    result = response.json()
-    if "result" not in result or "word" not in result["result"]:
-        return Response(
-            status=500,
-            result=None,
-            error=ErrorInfo(
-                code=500, message="Unexpected response format from Yahoo API"
-            ),
-        )
-
-    words = result["result"]["word"]
-    parsed_result: list[WordResult] = []
-
-    for word in words:
-        if "subword" in word:
-            subword_list = [
-                WordResult(furigana=sub["furigana"], surface=sub["surface"])
-                for sub in word["subword"]
-            ]
-            parsed_result.append(
-                WordResult(
-                    surface=word["surface"],
-                    furigana=word["furigana"],
-                    subword=subword_list,
-                )
-            )
-        else:
-            parsed_result.append(
-                WordResult(
-                    surface=word["surface"],
-                    furigana=word.get("furigana", word["surface"]),
-                )
-            )
-
-    return Response(status=200, result=parsed_result)
